requirements.txt

amphoradata==0.10.14
requests
python-dotenv